eom_tool: Add support to print EOM data JSON to stdout
Add support to print EOM data to stdout to meet requirements from
the Margin tool. The '-f' argument is now optional. When no output
file is specified, the tool outputs EOM data in JSON format to stdout
only. When '-f' is provided, the data is written to the specified file.

Change-Id: I2cc15a52b5ee9786d79e4808afee5b879921e71a
Signed-off-by: Arnav Kumar Dawn <arnavkum@qti.qualcomm.com>